New Windows, New Code. Getting Better

+++ src/main/java/ir/tic/clouddc/center/CenterRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CenterRepository extends JpaRepository<Center, Short> {
+public interface CenterRepository extends JpaRepository<Center, Integer> {
     List<Center> findAllByIdIn(List<Long> ids);
 
     @Query("SELECT C.id, C.name FROM Center C")
