+++ src/main/java/ir/tic/clouddc/pm/PmCategory.java
@@ -15,7 +15,7 @@ public class PmCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private short id;
+    private int id;
 
     @Column
     @Nationalized
@@ -29,11 +29,11 @@ public class PmCategory {
     private Persistence persistence;
 
 
-    public short getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(short id) {
+    public void setId(int id) {
         this.id = id;
     }
