+++ src/main/java/ir/tic/clouddc/pm/PmInterface.java
@@ -15,7 +15,7 @@ public final class PmInterface {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private short id;
+    private int id;
 
     @Column
     @Nationalized
@@ -67,11 +67,11 @@ public final class PmInterface {
         this.description = description;
     }
 
-    public short getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(short id) {
+    public void setId(int id) {
         this.id = id;
     }
